ongoingclass event into individual file
